chore: separate component definitions from component server
commit-id:56f63265
