+++ crates/mempool_infra/src/component_definitions.rs
@@ -1,0 +1,16 @@
+use async_trait::async_trait;
+use tokio::sync::mpsc::Sender;
+
+#[async_trait]
+pub trait ComponentRequestHandler<Request, Response> {
+    async fn handle_request(&mut self, request: Request) -> Response;
+}
+
+pub struct ComponentRequestAndResponseSender<Request, Response>
+where
+    Request: Send + Sync,
+    Response: Send + Sync,
+{
+    pub request: Request,
+    pub tx: Sender<Response>,
+}
