+++ crates/mempool_infra/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod component_client;
 pub mod component_client_rpc;
+pub mod component_definitions;
 pub mod component_runner;
 pub mod component_server;
 pub mod network_component;
